Added the code to handle service_reply_pkt to 32/64 bit clients.
The SERVICE protocol is basically sketched out and now I need to test it.
A big thing was the protocol API functions now do the host/network byte
ordering so I had to rip all of the explicit calls of them out of the
SERVICE protocol path in the server code. I will do the same for the
other protocols as I get to them.

I've done basic testing on a 32 bit platform, but I also need to test
the 64 bit code paths as well.  I can't do the cross platform test yet
until the STORE and RESTORE protocols are completed.

I also fixed the shadow to get rid of a stupid debug output--the
infamous "Weird 0xaabbccdd" line in the logs, which is just confusing
and unnecessary.