Lift Submitter/Scheduler MyType hack out of ClassAds.
As an old compatibility hack, submitter ads sent to the collector
by the schedd have a MyType of "Scheduler", but
ClassAd::GetMyTypeName() would return "Submitter" if the ad really
is a submitter ad. How does it distinguish between submitter ads
and real scheduler ad? Scheduler ads have a NumUsers attribute and
submitter ads don't.

This hack has now been lifted out of the ClassAd library, while
maintaining compatibility. Submitter ads now have a MyType of
"Submitter" (the collector enforces this, as does the schedd).
New daemons and tools are still able to deal with the old-style
submitter ads.

All of the compatibility bits are labeled with 'CRUFT' and should
be removed somewhere down the line.