+ Fixed a time consuming bug where I was miscalculating a pointer position in writing the additional packet data for a 64 bit service_req_pkt into a buffer.
+ Fixed a bunch of log messages and EXCEPT cases that could just be fail cases.

+ "Fixed" the uint64 byte ordering functions to be broken in the exact way I
need them to work. The exact detail of the sins that will lead me to hell are
detailed in a comment by the offending code. This is a result of C's types
changing over time and some ossification of it into an on the wire network
protocol.

+ The SERVICE protocol now seems to be at least as good as it was before I
added 1000 lines of backwards compatibility code to it. I haven't tested it
across 32/64 bit domain yet, as I need to do all the protocols before
I can do that due to their intertwined nature.

+ Ticket #177