Resource weights were not correctly handling sorting of matches.
GroupQuotaPermits() was incrementing the "used" tally and it was being
called while building the list of matches (i.e. before sorting takes place).
The result is that preferred matches would be ignored if they happened to
be found after the "used" tally exceeded the limit.

It is also necessary to invalidate the MatchList cache if it somehow
survives across pie spins and the user's usage limit is larger in the
subsequent spin of the pie (expected to happen if others drop out because
they don't have enough matching jobs). #331