Implemented a blocking connect with timeout in the schedd to ckpt server. This is to stop the schedd from blocking for long periods of time on down ckpt servers and getting itself into trouble. The schedd will remember timed out ckpt servers and not connect to them until some configurable time later.
This happens to cause *lots* of leaked checkpoints into the checkpoint server
under certain reasonably common conditions, so I also implemented a facility
for the checkpoint server to walk its CKPT_SERVER_DIR and evict files which are
considered stale according to some newly added knobs. No more stupid cron jobs,
it'll just do the right thing now--and it is off by default, to maintain
current behavior. The ckpt server will block while doing the cleaning
because it was easier to implement. If it becomes a problem, we can shove
it off into its own child process.

A more real solution would have been the schedd to defer and retry the removal
when the ckpt server came back up, but frankly, the schedd is complicated
enough. Also, the ckpt server could have called back to the schedd asking
about various checkpoints it thinks are stale, but that also would add more
complexity to the schedd and the schedd is busy enough without lots of
checkpoint servers (in the case of one per execute machine) deciding to talk
to it suddenly (plus the security ramifications of allowing the ckpt server
connection back to the submit machine).

I also converted the server_interface.c file to a cpp file so I could use the
STL with blocking connect with timeout work. It made my life a lot easier.

This is for ticket #552.