Allow cleanup of stranded VM jobs after a crash.
The starter can now write a recovery ad to disk with information about
the job it's running. For now, this is used to kill VMs that may have
been leaked during a crash of the Condor daemons.
In the future, this could be used to resume handling of jobs across a
planned restart of the startd and its children (as part of an upgrade).
gittrac #514