Re-check VM universe capability in startd after it is disabled. gittrac #551
Now, whenever the startd's check for whether the VM universe works
returns failure, the startd will periodically rerun the check until it
succeeds or VM_TYPE is not defined in the config file. The latter case is
an indication that the admin doesn't expect the VM universe to work.

When the VM universe transitions from not working to working, the slot
ads aren't updated immediately, but only on the next periodic update.