Initial support for GRAM5, user has to ID servers. (#675, #676)
When the user IDs a server as GRAM5 (via grid-type 'gt5'), the
gridmanager will ignore the running jobmanager throttle, disable the
grid monitor, and no explicitly stop the jobmanagers. The first job seen
for a resource determines whether the resource is treated as GRAM2 or
GRAM5.

Eventually, the gridmanager will autodetect whether a server is GRAM2 or
GRAM5, and grid-type 'gt5' will either disappear or be treated
identically to 'gt2'.