Fix curl external on hpux and freebsd.
The fix for solaris broke hpux and freebsd, but the externals caching
masked it. This is one of the dangers of changing an external build
script without changing the version.

This change is what should have been done. It fixes the problem on
solaris without changing things on any other platforms.