Allow cream external to build on x86_64 rhel 5.
The x86_64 rhel 5 build started failing because boost was depending on
the bzip2 library, which isn't on our nmi build machine for this
platform. I don't know why it built successfully before.

A check of the boost libraries and headers we install shows that this
change won't affect them, so it's safe to make without bumping the
version number.