Revert "Fixed 3-way deadlock observed in windows job_core_perhold_van. #735"
This reverts commit 0c30d66a8c997a0bd39999c008082b1d556ac61d.

This patch broke job_core_perremove-van.  Apparently, the call to
check periodic expressions in the update from the starter is depended
on for proper timing of this operation during job termination.  With
the delay in evaluation, periodic_remove was not firing at job exit time
when it was expected to.