Ignore GRID_MONITOR_NO_STATUS_TIMEOUT for unsubmitted jobs. #773
In the gridmanager, when processing a grid monitor job status report,
GlobusResource signals all jobs that haven't had a remote job status
update in the past GRID_MONITOR_NO_STATUS_TIMEOUT seconds. This is
meaningless for unsubmitted jobs and the corresponding GlobusJob
objects ignore the signal. But the signalling takes up a lot of time
when there are tens of thousands of jobs.

Now, GlobusResource doesn't signal jobs that don't have a remote job
id.