Fix for disappearing GT4 file transfer directory. #460
The empty directory needed for GT4 GRAM file transfers is disappearing
occassionally on some sites. We don't know what's removing it, but now
we'll occassionally check for its existence and recreate it if
necessary. We do this check at most once a minute, triggered when a
new job is received by the gridmanager or the gridmanager starts a
job submission.