Fixed spurious spool dir creation and deletion in the schedd.
In the past, we have created and deleted sandboxes for submissions
that do not require sandbox directories (typically, those that do not
require file transfer).  This is a problem in environments using NFS
spool directories (or spool directories on other shared filesystems),
since these operations are between one and two orders of magnitude
slower on NFS (and may also unacceptably load the NFS server).

We ensure that job sandboxes are never created or inspected by the
schedd for jobs whose NeverCreateJobSandbox attribute is set.

This is a partial fix for ticket #749 and RHT BZ 523806; subsequent
patches will correct spurious-spool behavior in the shadow and set
NeverCreateJobSandbox at submission time for a large class of jobs
that do not require sandboxes.