Changed default MAX_JOBS_RUNNING from 200 to something more appropriate. #767
The new default is min($(DETECTED_MEMORY)*0.8*1024/800,10000).
In Windows, we additionally cap this at 200.

  0.8 is our guess at how much of the memory it is ok to allocate to shadows

  800k is our guess at how much private memory each shadow needs
   (32-bit linux shadows with default configuration seem to need more like 400k
   and 64-bit linux shadows seem to need about 800k)

  10000 is our estimate of how many shadows we can run before risking
   exhaustion of ephemeral ports.  If the admin increases the ephemeral
   port range, it is possible to scale higher (up to 23000 in my experience).

This configuration change is being made at the param call-site for 7.4.
In 7.5, it will have to be done as an expression in the param table.

In choosing this default, we are ignoring the fact that scheduler
and local universe jobs are included in the limit.  We will limit
those universes further using the schedd start expressions.  We
expect those limits to be small compared to MAX_JOBS_RUNNING, so
their inclusion in the total running job count will typically not
lead to starvation of other universes.