Fixed memory leak in the collector introduced in 7.3.2.
The problem appears to have been from nearly simultaneous checkins
93d42438f (Dan) and 82f276c8 (Matt).  The former introduced a new
reference to USE_STRING_SPACE_IN_CLASSADS and the latter removed
the definition of USE_STRING_SPACE_IN_CLASSADS and all uses of it
_except_ for the new one that had just been added.  This resulted in
the wrong code path being followed, and the dangerous semantics of
the Variable constructor struck again.  Fortunately, that is now
a thing of the past and we will (I hope) not have any more forward
merges of code that relies on USE_STRING_SPACE_IN_CLASSADS.

Lesson: it is dangerous to remove macros used for conditional compilation.
They should probably be removed at the end of a development series.