Replace buggy GetAttributeExpr with GetAttributeExprNew().
GetAttributeExpr() would return either the attribute value as
a string or the string <name>=<value>, depending on whether the
attribute had been modified in the current queue-management
transaction. The new function always returns just the attribute value.

GetAttributeExpr() takes a pointer to a static buffer that it
writes into, leading to buffer overflows. The remote version won't
work with the underlying CEDAR calls to receive a string. The new
function works like GetAttributeStringNew(), allocating a buffer of
the right size and returning a pointer to the caller.

The underlying wire protocol remains the same.