Fix calling semantics of GetAttributeStringNew().
GetAttributeStringNew() allocates and returns an empty buffer on
failure most of the time. Allocating a buffer on failure is
surprising and some callers aren't expecting it. Only doing it for
some types of failures is horrible. Now, GetAttributeStringNew()
always sets the value pointer to NULL on failure.

Also fix a couple call sites that didn't know 1 can be returned on
success.