Provided a default amount of swap space for dynamic slots
A potentially better fix would be some calculation taking into account
the number of Cpus and the amount of requested memory. However, the
utility of VirtualMemory for policy is questionable.

VERSION HISTORY: Previously the VirtualMemory on dynamic slots was a multiple of the total virtual memory, sometimes becoming negative. It is now a fixed, and unmeaningful, percent. It should not be used in policy expressions. TotalVirtualMemory can still be meaningfully used in policy expressions.