Fixed signature of SpawnPreScript, updated usage of SpawnJob, #946
Notes: In all cases except for the timer case, a failure of SpawnJob resulted in shutdown, now all cases result in shutdown. It is debatable if SpawnJob should do the shutdown instead of each of the call sites. SpawnJob already does shutdown for some failure cases, so all code that calls it already must handle initiating a shutdown.