Fix ClassAdList handling in negotiator for compat layer. #619
ClassAdList in old ClassAds has reference counting. When an ad is
inserted into multiple lists, it is not deleted until it's removed from
all of the lists. ClassAdList in the compatibility layer for new
ClassAds doesn't have this, so the caller must determine when to delete
an ad that's in multiple lists.

Here, we modify the matchmaking code in the negotiator to work with both
old and compat ClassAdLists. Old ClassAdList gains a Remove() function
that's an alias for Delete(). In compat ClassAdList, Remove() removes
the indicated ad from the list without deleting it. In the negotiator,
all of the startd and submitter ads are in two ClassAdLists, requiring
the careful handling.