Improve handling of connection errors for cream jobs. #955
Cream gahp commands that fail due to connection-related problems are now
retried. If they continue to fail, a ping of the resource is attempted
to see if it is down. The number of retries and interval between them
can be controlled with params GRIDMANAGER_CONNECT_FAILURE_RETRY_COUNT
and GRIDMANAGER_CONNECT_FAILURE_RETRY_INTERVAL.