In vivo test of changed NMI glue for #715.
Due to the nastiness detailed in #974, I have to squash merge the
V7_4-submit_info-branch into the master to test to see if the new glue changes
I have made work for nightlies, continuous builds, and workspaces. I have
done what testing I am able given the contraints mentioned in #974 and this
is a pretty educated guess that it will function properly.

Regardless of the outcome of this merge: the builds fail miserably or pass;
it is intended that this commit be reverted by noon on Friday Nov 20th, 2009.