Got rid of TCP socket cache in collector. #997 It's only purpose was to restrict the number of registered sockets. DaemonCore does a better job of that, because it takes into account the file descriptor table size and is aware of other sources of registered sockets such as CCB.
This means it is no longer necessary to specially configure the collector
to accept TCP updates, other than making sure MAX_FILE_DESCRIPTORS is
big enough.

A proper authorization check is now done on the subsequent commands
received on the TCP update socket.  Previously, it was assumed that
all update commands had the same authorization policy, which is
not necessarily the case.