Always have a name to represent unauthenticated users. #1131 The name is unauthenticated@unmapped. This makes it possible to refer explicitly to unauthenticated users in the authorization policy.
This also changes behavior when both parties configure authentication
as something other than REQUIRED.  In this case, if authentication
is tried but fails, condor used to reject the connection.  Now it does not.
It is up to the authorization policy to determine whether the
unauthenticated user is allowed or denied.

Users who were authenticated but not mapped used to be mapped
to <method>@unmappeduser (e.g. gsi@unmappeduser).  They are now
mapped to <method>@unmapped.

Unmapped users used to be allowed to do commands that always require
authentication regardless of configuration (such as submitting jobs).
Anyone in the unmapped domain is no longer allowed to do such commands.