Force a TARGET scoping check in condor_submit. #1118
ClassAd::AddTargetRefs() does a version check on the ad and returns
immediately if the CondorVersion attribute indicates 7.5.1 or higher.
This assumes new Condor code does the scoping check before inserting
expressions into ads. Condor_submit works a little differently, building
the job ad, then checking it. So we need to force a check in that case.