Added the ability to specify a communal description and platform list to condor_nmi_submit when it submits the various nightly tags. This is so we can submit a different set of nightly builds than our one for the official ports.
One caveat to understand is that only the origin/master's copy of this
script is the one used for the nightlies, and subtley, if you make
a change to this script, it will only update to become live AFTER a
nightly.pl invocation. So, if you wait until the nightlies run, it'll
be missed that night, but active the next night. Therefore, you must
manually update the file in ~cndrauto/condor by running the git command by
hand found in the script for it to be active on the night you expect.

This is for ticket #1199.