For performance, made DestroyProc use a NONDURABLE transaction. #1046
The down side: by not immediately syncing the destroy proc event to
the job queue log, we widen the window of time in which a system crash
could cause the restarted schedd to redo job cleanup tasks that have already
been done (e.g. writing a history entry for the job).  However, we are only
widening the window, not creating it.  A system crash that happens during
job cleanup, before the writing of the destroy proc to the job queue log can
already result in repetition of job cleanup tasks.

The up side: removal of 100k jobs requires one fsync (to set them all to
X state) rather than 1+100k fsyncs.  Peak possible job throughput rate should
also be higher in the absence of this fsync.