Add implicit target scoping to new ClassAds. #1234
When evaluating an expression, if an unscoped attribute reference can't
be found in the current scope, allow a secondary scope to be searched as
well.

This is a temporary ability added to ease Condor's transition from old
to new ClassAds. In a MatchClassAd, each ad's alternate scope would be
the other ad. Once the transition is complete, we expect to remove this
change.