Add knob to disable implicit target scoping in compat ClassAds. #1242
Now that we can have implicit target scoping of attribute references in
new ClassAds (#1234), we should have a way to disable it. This will
allow users to test their configuration before upgrading to a future
version of Condor where we remove the old-ClassAd-style scoping.

The knob is called STRICT_CLASSAD_EVALUATION and it defaults to False.