Fixed problem of multiple submit events in event log. My recent change that moved CloseConnection() into CommitTransaction() (#1231) caused some references to old_cluster_num to be made when no qmgmt session was active, and the value of this global variable was not properly maintained outside of a qmgmt session, resulting in multiple submit events for the same job in the global event log.
The resolution was to get rid of this global variable and replace it with
an explicit call to the classad log object to get a list of new ads created
in a transaction.  The old mechanism for detecting new ads seems suspect in
the case where multiple qmgmt sessions are active at the same time
(i.e. with SOAP).  The new mechanism should be correct in this case.