Fix crash in gt2 gahp caused by jobmanager version command. #1279
The new gram client call globus_gram_client_get_jobmanager_version()
returns a valid hashtable on some errors, but not others. The gahp
assumed it always returns a valid hashtable. If the hashtable is valid,
then the gahp must free it to avoid leaking memory.

Now, the gahp initializes the hashtable pointer to NULL. If the pointer
is modified by globus_gram_client_get_jobmanager_version(), the gahp
frees it. Also, the gahp returns the contents of the hashtable only
when globus_gram_client_get_jobmanager_version() returns success.