Initial attempt to add support for dynamic gridftp server for cream jobs. #1400
Previously, the dynamic grdiftp server we start up for gt4 gram jobs
couldn't be used for cream jobs because the cream api had no way to
specify what DN the transferer should expect from the gridftp server.
Only servers with a host-based certificate could be used.

The most recent version of cream adds support for this by allowing a
client to include the server DN in the URLs of files to be transfered,
like so: <url>?<dn>

This is an initial test of using that feature. The gridmanager now
assumes all cream servers support this new feature.