Dynamic directory names no longer reappended on reconfig. #92
If a daemon is given -d on the command line, it modified the values of
LOG, SPOOL, and EXECUTE by appending its IP address and pid. It sets
the appropriate environment variables so that its children will use the
modified values. This is used for glide-in, so that multiple glide-ins
going to the same machine different different log, spool, and execute
directories.

Before this commit, each time the daemon received a reconfig, it appended
another copy of the IP address and pid to the modified values of LOG, SPOOL,
and EXECUTE. Now, it keeps the old modified values.

One side-effect of this commit is that changes to LOG, SPOOL, or EXECUTE
in the config file will be ignored on a reconfig when -d is used. Without
more extensive changes, any child daemons ignore any changes to these
parameters. Having the daemon keep the same values as its children
seems the best way to go.