Revert "Send RELEASE_CLAIM for match_recs that are in M_STARTD_CONTACT_LIMBO (#742)"
This reverts commit c6ef74d2c2da9becaa275d64d8cccf471871c000.

While in M_STARTD_CONTACT_LIMBO, we do not want to send RELEASE_CLAIM to
the startd or we may delete a claim that some other schedd has a prior
claim to (e.g. if the negotiator has stale info and handed out the same
claim in different negotiation cycles).