Fix idle scheduler jobs crashing schedd when spooling input files. #1330
When a job's input files are transferred to the schedd over the network
as part of a job submission, the job is submitted on hold, the files are
transferred, then the job is released by the schedd. The function in the
schedd that's called once the transfer is complete
(Scheduler::spoolJobFilesReaper()) does the following:

* Start a queue transaction
* Modify some attributes in the job ad
* call releaseJob()
* Commit the transaction

releaseJob() ends up calling Scheduler::needReschedule(), which calls
Scheduler::timeout(), which will attempt to start any idle scheduler
universe jobs, using a transaction to update the job ad. So if there are
any idle scheduler universe jobs, the schedd attempts to start two
transactions in parallel, which blows an ASSERT().

Now, needReschedule() just reset's timeout()'s daemon-core timer to go
off immediately, rather than calling timeout() directly. This moves the
call to timeout() out of the transaction of spoolJobFilesReaper().
timeout() can already decide to delay the next time it runs, so this
change shouldn't break any assumptions made by callers to
needReschedule().