Disable Amazon GAHP build on platforms without threads. #1362
The Amazon EC2 GAHP uses threads and links with the Condor libraries.
This will lead to undefined behavior on platforms where we don't build
the Condor libraries with thread support. The Amazon EC2 test is failing
on Solaris because of an incorrect errno value.

Linux is the only platform where we build the Condor libraries with
thread support (i.e. use -pthread), so the Amazon GAHP should only be
built on Linux.