Make vmware_dir submit attribute optional for vmware jobs. #1351
If vmware_dir isn't specified in the submit file, all of the VM image
files must be listed in transfer_input_files. condor_submit will check
that list and complain if a .vmx file isn't present. The job ad
attributes VMPARAM_VMware_VMX_File and VMPARAM_VMware_VMDK_Files are
set the same as before. VMPARAM_VMware_Dir is now optional in the job
ad.

This allows VM image files to be pre-staged on execute machines (using
file transfer plugins) and not be visible on the submit machine.