Allow new ClassAds to convert numerics to booleans in expressions. #1203
Expressions such as '0 || 1' are illegal in new ClassAds but legal in
old ClassAds. We have some users who use this coercion from numeric to
boolean values.

To ease the transition from old to new ClassAds in Condor, we are adding
a flag to allow this evaluation behavior in new ClassAds. It's disabled
by default and is meant only for use with Condor.