Allow conversions of numerics to booleans in ClassAd compat layer. #1203
Expressions such as '0 || 1' are illegal in new ClassAds but legal in
old ClassAds. We have some users who use this coercion from numeric to
boolean values.

In the ClassAd compatibility layer, take advantage of a new flag in new
ClassAds which allows this coercion during expression evaluation. The
compatibility layer enables the flag be default, but it can be disabled
by setting STRICT_CLASSAD_EVALUATION=True in the config file.