Fix nordugrid_gahp portability problems due to ldap dependencies. #1383
The nordugrid_gahp is linked against the system OpenLDAP library. On
many of our linux platforms, OpenLDAP is dependent on SASL or SASL2. For
static linking, the SASL/SASL2 library needs to be included on the link
line. With the way our build scripts are structured, the same link
options are used for dynamic and static.

It turns out that explicitly naming SASL/SASL2 on the link line makes
the binary less portable between linux flavors. The WMS group discovered
that a dynamic nordugrid_gahp built on rhel3 seg-faults on scientific
linux 4. This is because rhel3's ldap uses sasl, while sl4's ldap uses
sasl2. Both sasl libraries end up in the running process and butt heads.
Removing the sasl library from the link line fixes the problem, but then
the static link fails.

For now, I'm hacking the makefile to use different lists of libraries
for the dynamic and static links of nordugrid_gahp. This requires
manually expanding an imake macro, which is ugly and wrong. If we can
dump the building of static binaries (#1136), this mess will go away.