Make EMPTY_DESCRIP constant in daemon-core. #25
EMPTY_DESCRIP is a default description string for daemon-core's
Register_*() methods. It's assigned a constant value but is declared
non-constant so that it doesn't have to be strdup()d every time a caller
passes NULL to one of the Register_*() methods. This causes the compiler
to rightfully complain.

I checked a couple daemons and they never pass NULL to the Register_*()
methods, so this time/space-saving optimization has little or no
benefit. So now EMPTY_DESCRIP is changed to be constant and is strdup()d
anytime it's used.