Fix partial line reads in the Deltacloud GAHP.
The PipeBuffer code that we had been using had a problem
in that it wouldn't properly concatenate partial lines
into a single complete line.  Because of this it would
sometimes only get half of a command, and confuse the upper
levels.  After looking at it I decided that for the Dcloud
GAHP we needed much simpler code, so I mostly gutted the
PipeBuffer class and made it do the one simple thing we
need.