This checkin implements #1195. The new script run_nightly_builds now reads the nightly tags file, checks out the correct nmi_tools associated with a nightly tag, and submits the build using that version of nmi_tools.
--nightly in condor_nmi_submit used to do the above behavior, but for
reasons in #1195 and #974 this turned out to not be too functional
when the build glue needs a radical change on a branch.

This single commit combines the new nightly script and the condor_nmi_submit
build changes because it turns out they need to know a little bit about how
the other works.

If you're reading this because the nightlies blew up and I'm not around to
fix it and you absolutely need it to work:

1. Revert this commit on V7_4-branch, merge to master, push to central
	repository, git fetch the nmi-s006 clone.
2. Change the cndrauto crontab on nmi-s006 to use the old method.
3. In nmi-s006:/home/cndrauto/condor/old, type
	mv * ..

YMMV, but this will restore the older system.