Allow user-defined new ClassAds functions via CLASSAD_USER_LIBS. #1228
Add new Condor config param CLASSAD_USER_LIBS to allow the user to
provide shared libraries with user-defined functions for new ClassAds.
The old parameter CLASSAD_LIB_PATH is no longer supported, since the ABI
is different between new and old ClassAds. A list of libraries can be
provided. classad::FunctionCall::RegisterSharedLibraryFunctions() is
called for each one. A list is kept to allow additional libraries to be
added on a reconfig, but avoid re-registering the same libraries twice.