Bad ldap server can make nordugrid_gahp exit. #1404
The nordugrid_gahp exits when it receives a SIGPIPE. The assumption is
that a SIGPIPE indicates that one of the pipes to its client has been
closed. However, a NORDUGRID_LDAP_QUERY command issued to a
misconfigured ldap server can also result in a SIGPIPE. This is because
a SIGPIPE is generated when the process writes to a closed pipe or
socket.

Igor Sfiligoi had his nordugrid_gahp exit unexpectedly when attempting
to query an ldap server that closed connections immediately after they
were opened. The connections stayed open long enough for the ldap
library to start sending data, triggering a SIGPIPE.

The solution is for the nordugrid_gahp to ignore SIGPIPE and exit if
writing to stdout fails with EPIPE. Fortunately, virtually all writes to
stdout are funneled through gahp_printf(), which makes adding a check
easy.