Control file staging in the gridmanager for cream jobs. #1403
This is the first draft for initiating file staging from the client side
for cream grid jobs, rather than having the server perform the
transfers.

The gridmanager has a new TransferRequest class which will manage the
transfer of a list of files. It uses the new GRIDFTP_TRANSFER command of
the nordugrid_gahp. Eventually, TransferRequest can be used for other
job types (gt2, gt5, nordugrid) and limit the number of concurrent
transfers across all jobs. For now, the only limit is
GRIDMANAGER_MAX_PENDING_REQUESTS.