Extend the lifetime of a gahp server by a short period.
When the last GahpClient object for a gahp server is deleted in the
gridmanager, the GahpServer object and corresponding gahp server process
are kept around for 30 seconds before being destroyed.

With the new TransferRequest class, we may otherwise start to see periods
of repeated launching and killing of gahps as TransferRequest objects are
created and deleted.