Fix ClassAd function interval() to better match documentation. #1440
In breaking down a time interval into days, hours, minutes, and seconds,
the interval() function always space-padded the days to three characters
and never omitted leading zeros, both of which conflict with the
documenation. These behaviors have been changed to match the docs.

In a slight departure from the docs, the hours value is zero-padded to 2
characters when the days value is non-zero. I believe the docs should be
changed to match this. Also, negative intervals are now displayed in a
sane fashion.