Fix bugs in daemon log lock file code that can cause EXCEPT()s. #1484
When multiple daemons writing to the same daemon log are using a lock
file (i.e. <subsys>_LOCK is defined), a daemon rotating the log will
EXCEPT if the log filename still exists immediately after the daemon
renames it (since the lock file should prevent any other daemons from
writing to it during the rotation). This commit fixes two cases where
the lock file didn't prevent another daemon from writing to the log
during rotation.

The first instance occurs if the lock file is deleted. Daemons open the
lock file once and keep it open until they exit. If the file is removed,
the next daemon to start will create a new lock file, resulting in two
lock files, each used by some of the daemons. Thus, exclusive access to
the log is lost. The fix is to have each daemon fstat() the lock file fd
each time it wants to lock it. If the link count is 0, the daemon closes
and reopens the file. There is still a window of vulnerability, but it
is now tiny.

The second instance occurs when a daemon first starts. When the daemon
logging code is initialized, the log file is opened and closed to ensure
it's accessible. This creates the log file if it doesn't already exist.
This is done without regard to the lock file. The code then checks
whether a lock file is configured. Now, the lock file configuration is
checked first, and the initial open/close of the log file is protected
by locking the lock file.