Added support for directory trees in Condor's file transfer mode. #1522
This does NOT cause auto-selection of output to look in subdirectories.

This patch also happens to fix a problem with spooling of output. #1521
Previously, if output was auto-selected, condor_transfer_data would
grab both the spooled output and the spooled input files.
Also (#1520), if files in the transfer output files list were specified
with paths and not simple basenames, condor_transfer_data would fail
to fetch the spooled output, because it tried to read from the original
path rather than the basename.