Allow semicolon as argument separator in old ClassAd emulation mode. #1516
In old ClassAds, arguments to a function can be separated with commas or
semicolons. In new ClassAds, only commas can be used. In old ClassAds,
the use of semicolons is an older syntax that is now avoided, but still
valid.

We've identified two users of Condor who use the semicolon notation.
They are the Swift workload management system
(https://svn.ci.uchicago.edu/svn/vdl2/usertools/swift/osgedu-list-sites)
and the ReSS service of OSG. As such, we should make new ClassAds
recognize this notation when in old ClassAd compatibility mode (i.e.
when _useOldClassAdSemantics is set to true).