Fix escaping of error messages in nordugrid gahp #1537
When passing error messages from the gridftp library back to the
gridmanager, the nordugrid gahp converts newlines to spaces. It does
this because the error messages often end up in the HoldReason
attribute of the job ad, and the schedd can't handle newlines in
attribute values. But the gahp doesn't escape these spaces, which makes
the gridmanager unable to parse the gahp's response, causing the
gridmanager to EXCEPT. The escaping is now done properly.