Add config param SOAP_SSL_SKIP_HOST_CHECK #1545
Setting this parameter to True causes the amazon_gahp to set the gsoap
flag SOAP_SSL_SKIP_HOST_CHECK when connecting to servers. This disables
the check for the server's hostname matching that in its certificate.
The default is False (don't disable the check).

This is needed for some Eucalyptus servers, which use a certificate
which contains a hostname of localhost.