remove name format checks in amazon_gahp #1545
The amazon_gahp used to check whether image and instance ids match a
particular pattern that Amazon's EC2 service is known to adhere to.
Specifically, that image ids start with "ami-" and instance ids start
with "i-". These patterns aren't part of the EC2 specificatoin, and
Eucalyptus doesn't follow them. The checks are now gone.