Fix parsing of negative nubmers on new ClassAds #1544
The new ClassAds parser parses -1 as '-' '1'. That is, the unary minus
operator followed by the integer 1. This works ok for most integer and
real values, as when the expression is evaluated, the correct value is
reconstructed. INT_MIN (aka -2147483648) is an exception, since the
negated value (2147483648) can't be represented in signed 32-bits. The
parser calls strtol() on the string representing the number (without the
minus), then casts the result to an int. On a 32-bit machine, strtol()
detects overflow and returns LONG_MAX (which is the same as INT_MAX,
2147483647). On a 64-bit machine, 2147483648 is returned in 64 bits,
which is then truncated to 32 bits by the cast, resulting in the value
-2147483648. When this is unparsed, an extra minus sign ends up in front
(for the unary minus operator). If the resulting string is parsed and
unparsed again, another minus sign appears.

Now, negative numbers are parsed as such, including INT_MIN. This
requires a little more knowledge in the lexer when encountering a minus
sign. It must consider the previous token and the subsequent 2
characters to decide whether the minus is the start of a negative
number.