Fixed the "unknown negotiator" problem. #1560
For better or worse, when flocking is enabled, the schedd needs to know
which flocking level it is dealing with when negotiating.  If it can't
figure it out, the schedd refuses to negotiate, with the message
"unknown negotiator".

The old algorithm required the admin to set NEGOTIATOR_HOST and/or
FLOCK_NEGOTIATOR_HOSTS to something with a DNS entry matching the
peer address observed by the schedd when receiving a connection from
the negotiator.

The new algorithm just works (when the negotiator is 7.5.4+).  Nothing
special has to be done for multi-homed negotiators or negotiators that
run on a different host from the collector.  The algorithm doesn't even
assume that there is only one negotiator per collector.

FLOCK_NEGOTIATOR_HOSTS is now optional in the schedd.  It's only use
is for identifying pre 7.5.4 negotiators and as a macro in the config
file for authorization policy.  If not set, the schedd simply queries
the address of the flocked negotiators (for sending RESCHEDULE).