Strip port from hostname for nordugrid ldap queries #1557
When performing an ldap query to a nordugrid resource, we need to strip
off any port given in the resource name in the job ad. Otherwise, the
ldap client will attempt to connect to that port, should will have a
gridftp server listening. The two dead-lock for ten minutes each waiting
for the other to say something.